inference_service/server.py
#!/usr/bin/env python3
"""
gRPC server for image classification inference.
"""

import grpc
from concurrent import futures
import logging

# TODO: Import generated protobuf classes
# import inference_pb2
# import inference_pb2_grpc

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


class InferenceServicer:
    """
    Implements the InferenceService defined in inference.proto
    """
    
    def Predict(self, request, context):
        """
        Handle inference prediction requests.
        
        Args:
            request: PredictRequest containing image_data
            context: gRPC context
            
        Returns:
            PredictResponse with predicted_class and confidence
        """
        logger.info("Received prediction request")
        # TODO: Implement actual model inference
        # For now, return a placeholder response
        return None


def serve(port=50051):
    """
    Start the gRPC server.
    
    Args:
        port: Port to listen on (default: 50051)
    """
    server = grpc.server(futures.ThreadPoolExecutor(max_workers=10))
    # TODO: Add servicer to server
    # inference_pb2_grpc.add_InferenceServiceServicer_to_server(
    #     InferenceServicer(), server
    # )
    server.add_insecure_port(f'[::]:{port}')
    server.start()
    logger.info(f"Server started on port {port}")
    server.wait_for_termination()


if __name__ == '__main__':
    serve()
