inference_service/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
torch==2.1.2
torchvision==0.16.2
pillow==10.1.0
numpy==1.26.2
